refactor: move RegisterUser tests to __tests__ folder
